⚡ Bolt: optimize BulkDeleteCustomers and standardize IDs
This commit implements a performance optimization for the bulk customer deletion
process and standardizes Customer ID types across the backend for consistency.

Key changes:
- Refactored `CustomerService.BulkDeleteCustomers` to resolve an N+1 bottleneck.
  - Replaced iterative individual fetches with a single batch fetch using `repo.GetByIDs`.
  - Parallelized external Shopify deletions using `errgroup` with a concurrency limit
    of 5 to respect API rate limits while significantly reducing total execution time.
  - Replaced iterative individual deletes with a single batch database delete using
    `repo.BulkDelete`.
- Standardized Customer ID types to `int64` across:
  - `CustomerRepository.GetByIDs`
  - `CustomerService.GetCustomersByIDs`
  - `AutomationHandler.SendBulkMarketing` request struct.
  This ensures alignment with the `entity.Customer` model and prevents potential
  mapping or overflow issues.
- Updated `go.mod` and `go.sum` to include `golang.org/x/sync`.

Impact:
- Reduces database roundtrips for fetching from O(N) to O(1) during bulk delete.
- Parallelizes external API calls to Shopify, slashing latency for bulk operations.
- Improves overall codebase consistency and type safety for customer-related operations.

+++ backend/internal/automation/whatsapp/handlers.go
@@ -947,8 +947,8 @@ func (h *AutomationHandler) SendBulkMarketing(w http.ResponseWriter, r *http.Req
 	}
 
 	var req struct {
-		CustomerIDs []uint `json:"customer_ids"`
-		TemplateID  int    `json:"template_id"`
+		CustomerIDs []int64 `json:"customer_ids"`
+		TemplateID  int     `json:"template_id"`
 	}
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
 		http.Error(w, "Invalid request body", http.StatusBadRequest)
